requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.0.0
pydantic>=1.8.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
alembic>=1.7.0
bcrypt>=3.2.0
python-dotenv>=0.19.0
requests>=2.25.0
minio>=7.1.0
boto3>=1.20.0
email-validator>=1.1.0
numpy>=1.21.0
httpx>=0.23.0
psutil>=5.8.0
pytest-asyncio>=0.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
google-generativeai>=0.7.0
safety>=3.0.0
pip-audit>=2.7.0
pytest-benchmark>=5.0.0
